tape: use u64 ptr -- I ought to be hanged for my legendary stupidity

+++ tape.h
@@ -8,7 +8,7 @@ struct Tape {
 	u8 *cells;
 
 	/* cell pointer */
-	u8 pointer;
+	u64 pointer;
 };
 
 #endif
